requirements.txt
backcall==0.2.0
Bottleneck 
certifi 
cffi 
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cryptography
cycler 
distlib==0.3.6
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
Flask 
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
fonttools==4.38.0
gunicorn==20.1.0
importlib-metadata
ipython-genutils==0.2.0
itsdangerous==2.1.2
Jinja2 
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe 
matplotlib==3.6.2
mistune==2.0.4
munkres==1.1.4
mysql==0.0.3
mysqlclient==2.1.1
numexpr 
numpy==1.23.4
packaging 
pandas==1.5.1
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.2
ply==3.11
prettytable==2.2.1
psycopg2==2.9.5
pure-eval==0.2.2
pycparser 
Pygments==2.13.0
PyMySQL 
pyparsing 
PyQt5==5.15.7
PyQt5-sip 
pyrsistent==0.19.2
python-dateutil 
pytz==2022.6
scikit-learn==0.24.2
scipy==1.9.3
Send2Trash==1.8.0
sip 
six 
sniffio==1.3.0
soupsieve==2.3.2.post1
tabulate==0.8.9
threadpoolctl==3.1.0
tinycss2==1.2.1
toml 
tornado 
tqdm==4.64.1
traitlets==5.5.0
virtualenv==20.16.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wincertstore==0.2
zipp 
